fix LEDs (I think)
The number of underglow LEDs is controlled by the chain-length property in the led_strip node. You can change the value of this property in the <keyboard>.keymap file by adding a stanza like this one outside of any other node (i.e. above or below the / node):

&led_strip {
    chain-length = <21>;
};

where the value is the total count of LEDs (per half, for split keyboards).